requirements.txt
dspy-ai>=2.0.4
openai>=1.3.0
typer>=0.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
rich>=13.7.0
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
ruff>=0.1.0
cloudinary>=1.36.0
pillow>=10.1.0
requests>=2.31.0
tenacity>=8.2.0
loguru>=0.7.0
locust>=2.24.0
psycopg2-binary>=2.9.9
fastapi>=0.109.0
uvicorn>=0.25.0
sqlalchemy>=2.0.27
alembic>=1.13.1
redis>=5.0.0
aioredis>=2.0.0
structlog>=24.1.0 